Create a new Conversation class
